+++ flutter_dex/lib/main.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dex/components/homeButtom.dart';
 import 'package:flutter_dex/components/bottomNavBar.dart';
 import 'package:flutter_dex/homepage/homePage.dart';
+import 'package:flutter_dex/quiz/telaQuiz.dart';
 
 void main() {
   runApp(MyApp());
@@ -26,50 +27,6 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class MyHomePage extends StatefulWidget {
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Dex basics'),
-      ),
-      body: const Center(
-        child: Text(
-          'Página Principal',
-          style: TextStyle(fontSize: 24),
-        ),
-      ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      floatingActionButton: DexButtomHome(),
-      bottomNavigationBar: DexAppBar(),
-    );
-  }
-}
-
-class TelaQuiz extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Dex basics'),
-      ),
-      body: const Center(
-        child: Text(
-          'Página quiz',
-          style: TextStyle(fontSize: 24),
-        ),
-      ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      floatingActionButton: DexButtomHome(),
-      bottomNavigationBar: DexAppBar(),
-    );
-  }
-}
 
 class TelaRegion extends StatelessWidget {
   @override
